cambiando las migraciones para que se puedan añadir decimales en los campos de cantidad de residuos en kg

+++ database/migrations/2019_03_29_082343_add_colums_solicitud_residuos_table.php
@@ -15,7 +15,7 @@ class AddColumsSolicitudResiduosTable extends Migration
     {
         Schema::table('solicitud_residuos', function (Blueprint $table) {
             $table->string('SolResTypeUnidad')->nullable();
-            $table->bigInteger('SolResCantiUnidad')->nullable();
+            $table->decimal('SolResCantiUnidad', 8, 2)->nullable();
             $table->string('SolResEmbalaje')->nullable();
             $table->bigInteger('SolResAlto')->nullable();
             $table->bigInteger('SolResAncho')->nullable();
